[BugFix]: Ticket #366080 -- Deleting an empty shift was STILL not working propertly.

+++ src/net/sourceforge/fenixedu/domain/Shift.java
@@ -161,7 +161,9 @@ public class Shift extends Shift_Base {
 
             getAssociatedClasses().clear();
             getCourseLoads().clear();
-            getShiftGroupingProperties().delete();
+            if (getShiftGroupingProperties() != null) {
+                getShiftGroupingProperties().delete();
+            }
             removeRootDomainObject();
             super.deleteDomainObject();
